NodeInputIcons as container
