+++ src/containers/Node/index.js
@@ -1,25 +1,18 @@
 import { connect } from 'react-redux'
 import Node from '../../components/Node'
 import getNode from '../../selectors/getNode'
-import { uSketchNodeOpenedToggle } from '../../store/sketches/actions'
 import getIsNodeOpen from '../../selectors/getIsNodeOpen'
 import getActiveInputsText from '../../selectors/getActiveInputsText'
 import { nodeShotFired } from '../../store/nodes/actions'
-import { uiNodeToggleOpen } from '../../store/ui/actions'
-import { rMacroOpenToggle } from '../../store/macros/actions'
 
 const mapStateToProps = (state, ownProps) => {
-  const node = getNode(state, ownProps.nodeId)
-  const inputLinkIds = node.inputLinkIds
-  const param = state.nodes[ownProps.nodeId]
+  const param = getNode(state, ownProps.nodeId)
   const type = ownProps.type || 'param'
 
   const inputLinkTitle = getActiveInputsText(state, ownProps.nodeId)
 
   return {
     type,
-    numInputs: inputLinkIds.length,
-    numMacros: node.connectedMacroIds.length,
     title: param.title,
     isOpen: getIsNodeOpen(state, ownProps.nodeId, ownProps.panelId),
     inputLinkTitle,
@@ -30,19 +23,6 @@ const mapDispatchToProps = (dispatch, ownProps) => {
   const type = ownProps.type || 'param'
 
   return {
-    onOpenClick: () => {
-      switch (ownProps.panelId) {
-        case 'overview':
-          dispatch(uiNodeToggleOpen(ownProps.nodeId))
-          break
-        case 'macros':
-          dispatch(rMacroOpenToggle(ownProps.nodeId))
-          break
-        case 'sketch':
-        default:
-          dispatch(uSketchNodeOpenedToggle(ownProps.nodeId))
-      }
-    },
     onParamBarClick: type === 'shot'
     ? () => {
       dispatch(nodeShotFired(ownProps.nodeId, ownProps.sketchId, ownProps.shotMethod))
@@ -50,9 +30,9 @@ const mapDispatchToProps = (dispatch, ownProps) => {
   }
 }
 
-const ParamContainer = connect(
+const NodeContainer = connect(
   mapStateToProps,
   mapDispatchToProps
 )(Node)
 
-export default ParamContainer
+export default NodeContainer
